+++ tests/Transformers/TfIdfTransformerTest.php
@@ -2,9 +2,11 @@
 
 namespace Rubix\ML\Tests\Transformers;
 
+use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Unlabeled;
 use Rubix\ML\Transformers\Elastic;
 use Rubix\ML\Transformers\Stateful;
+use Rubix\ML\Transformers\Reversible;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Transformers\TfIdfTransformer;
 use Rubix\ML\Exceptions\RuntimeException;
@@ -49,12 +51,14 @@ class TfIdfTransformerTest extends TestCase
         $this->assertInstanceOf(Transformer::class, $this->transformer);
         $this->assertInstanceOf(Stateful::class, $this->transformer);
         $this->assertInstanceOf(Elastic::class, $this->transformer);
+        $this->assertInstanceOf(Reversible::class, $this->transformer);
+        $this->assertInstanceOf(Persistable::class, $this->transformer);
     }
 
     /**
      * @test
      */
-    public function fitTransform() : void
+    public function fitTransformReverse() : void
     {
         $this->transformer->fit($this->dataset);
 
@@ -66,15 +70,21 @@ class TfIdfTransformerTest extends TestCase
         $this->assertCount(19, $dfs);
         $this->assertContainsOnly('int', $dfs);
 
+        $original = clone $this->dataset;
+
         $this->dataset->apply($this->transformer);
 
-        $outcome = [
+        $expected = [
             [1.6931471805599454, 3.8630462173553424, 0.0, 0.0, 1.2876820724517808, 0.0, 0.0, 0.0, 1.2876820724517808, 2.5753641449035616, 0.0, 2.5753641449035616, 0.0, 0.0, 0.0, 6.772588722239782, 1.2876820724517808, 0.0, 1.6931471805599454],
             [0.0, 1.2876820724517808, 1.6931471805599454, 0.0, 0.0, 2.5753641449035616, 1.6931471805599454, 0.0, 0.0, 0.0, 0.0, 3.8630462173553424, 0.0, 1.6931471805599454, 0.0, 0.0, 0.0, 0.0, 0.0],
             [0.0, 0.0, 0.0, 1.6931471805599454, 2.5753641449035616, 3.8630462173553424, 0.0, 0.0, 5.150728289807123, 2.5753641449035616, 0.0, 0.0, 1.6931471805599454, 0.0, 3.386294361119891, 0.0, 1.2876820724517808, 0.0, 0.0],
         ];
 
-        $this->assertEquals($outcome, $this->dataset->samples());
+        $this->assertEquals($expected, $this->dataset->samples());
+
+        $this->dataset->reverseApply($this->transformer);
+
+        $this->assertEquals($original->samples(), $this->dataset->samples());
     }
 
     /**